moved routes to app.js file for testing
